requirements.txt
pyTelegramBotAPI
python-dotenv
google-generativeai
